divide expand_asterisk for norm
